requirements.txt
hail==0.2.127
matplotlib==3.8.2
jinja2==3.1.2
openai==1.3.5
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
scikit-learn==1.3.2
pysam==0.22.0  # For handling VCF files
pyyaml==6.0.1  # For config file handling
tqdm==4.66.1   # For progress bars
prscs==0.1.0   # For PRS calculation
prscsx==0.1.0  # For cross-population PRS 